requirements.txt
chainlit
langchain
langchain-community
langchain-openai
python-dotenv
pypdf
qdrant-client